Updated win32 process class. Added process tree closing.

+++ StarfallAfterlifeBridge/Native/Windows/WindowsProcess.cs
@@ -1,4 +1,5 @@
 ﻿using Microsoft.Win32.SafeHandles;
+using StarfallAfterlife.Bridge.Diagnostics;
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
@@ -8,6 +9,7 @@ using System.Linq;
 using System.Reflection.Metadata;
 using System.Runtime.InteropServices;
 using System.Text;
+using System.Threading;
 using System.Threading.Tasks;
 using static StarfallAfterlife.Bridge.Native.Windows.Win32;
 
@@ -25,37 +27,59 @@ namespace StarfallAfterlife.Bridge.Native.Windows
 
         public int Id { get; private set; }
 
-        public IntPtr TheadHandle { get; private set; }
+        public IntPtr ThreadHandle { get; private set; }
 
-        public int TheadId { get; private set; }
+        public int ThreadId { get; private set; }
 
-        public bool IsStarted { get => isStarted; }
+        public bool IsStarted => _isStarted;
+
+        public bool AttachToMainProcess
+        {
+            get => _attachToMainProcess;
+            set
+            {
+                lock (_processLockher)
+                    if (_isStarted == false)
+                        _attachToMainProcess = value;
+            }
+        }
 
         public StreamReader StandartOutput { get; private set; }
 
-        private AnonymousPipeServerStream outputPipe;
-        private static readonly object startLockher = new();
-        private bool isStarted;
-        private bool disposedValue;
+        public bool HasExited => SharpProcess?.HasExited ?? true;
+
+        public nint MainWindowHandle => SharpProcess?.MainWindowHandle ?? nint.Zero;
+
+        public event EventHandler<EventArgs> Exited;
+
+        private AnonymousPipeServerStream _outputPipe;
+        private static readonly object _processLockher = new();
+        private bool _isStarted;
+        private bool _isExitCompleted;
+        private bool _disposedValue;
+        private ProcessWaitHandle _waitHandle;
+        private RegisteredWaitHandle _threadPoolWaitHandle;
+        private bool _attachToMainProcess;
+        private JobObject _job;
 
         public bool Start(string name, string cmd = null, StartupInfo startupInfo = default)
         {
-            lock (startLockher)
+            lock (_processLockher)
             {
-                if (isStarted == true)
+                if (_isStarted == true)
                     return false;
 
-                isStarted = true;
+                _isStarted = true;
 
-                if (disposedValue == true)
+                if (_disposedValue == true)
                     return false;
 
                 ProcessInfo processInfo = default;
 
                 Handle = IntPtr.Zero;
                 Id = -1;
-                TheadHandle = IntPtr.Zero;
-                TheadId = -1;
+                ThreadHandle = IntPtr.Zero;
+                ThreadId = -1;
 
                 try
                 {
@@ -64,8 +88,8 @@ namespace StarfallAfterlife.Bridge.Native.Windows
 
                     if (startupInfo.dwFlags.HasFlag(STARTF.USESTDHANDLES))
                     {
-                        outputPipe = new(PipeDirection.In, HandleInheritability.Inheritable);
-                        startupInfo.hStdOutput = outputPipe.ClientSafePipeHandle.DangerousGetHandle();
+                        _outputPipe = new(PipeDirection.In, HandleInheritability.Inheritable);
+                        startupInfo.hStdOutput = _outputPipe.ClientSafePipeHandle.DangerousGetHandle();
                     }
 
                     bool result = CreateProcess(
@@ -84,26 +108,36 @@ namespace StarfallAfterlife.Bridge.Native.Windows
                         processInfo.hProcess is 0 or -1 ||
                         processInfo.hThread is 0 or -1)
                     {
-                        CloseOutput();
+                        ReleaseUnmanagedResources();
                         return false;
                     }
 
                     Handle = processInfo.hProcess;
                     Id = processInfo.ProcessId;
-                    TheadHandle = processInfo.hThread;
-                    TheadId = processInfo.ThreadId;
+                    ThreadHandle = processInfo.hThread;
+                    ThreadId = processInfo.ThreadId;
                     SharpProcess = Process.GetProcessById(Id);
-                    SharpProcess.EnableRaisingEvents = true;
-                    SharpProcess.Exited += ProcessExited;
+                    StartWatchingForExit();
 
                     if (startupInfo.dwFlags.HasFlag(STARTF.USESTDHANDLES))
                     {
-                        StandartOutput = new(outputPipe, Console.OutputEncoding, true, 4096);
+                        StandartOutput = new(_outputPipe, Console.OutputEncoding, true, 4096);
                     }
+
+                    if (AttachToMainProcess == true)
+                    {
+                        _job = JobObject.Create(new()
+                        {
+                            LimitFlags = LimitFlags.KILL_ON_JOB_CLOSE,
+                        });
+                        
+                        _job.AttachToProcess(Handle);
+                    }
+
                 }
                 catch
                 {
-                    CloseOutput();
+                    ReleaseUnmanagedResources();
                     return false;
                 }
             }
@@ -111,27 +145,86 @@ namespace StarfallAfterlife.Bridge.Native.Windows
             return true;
         }
 
-        private void ProcessExited(object sender, EventArgs e)
+        private void ProcessExit()
         {
-            CloseOutput();
+            lock (_processLockher)
+            {
+                if (_isExitCompleted == true)
+                    return;
+
+                ReleaseUnmanagedResources();
+                Exited?.Invoke(this, EventArgs.Empty);
+            }
         }
 
-        private void CloseOutput()
+        private void ReleaseUnmanagedResources()
         {
-            outputPipe?.DisposeLocalCopyOfClientHandle();
-            outputPipe?.Dispose();
-            StandartOutput?.Dispose();
+            try
+            {
+                _outputPipe?.DisposeLocalCopyOfClientHandle();
+                _outputPipe?.Dispose();
+                StandartOutput?.Dispose();
+            }
+            catch { }
+
+            try
+            {
+                _job?.Dispose();
+                SharpProcess?.Dispose();
+            }
+            catch { }
+
+            try
+            {
+                _waitHandle?.Dispose();
+                _threadPoolWaitHandle?.Unregister(null);
+            }
+            catch { }
         }
 
         public void Kill()
         {
-            SharpProcess?.Kill();
+            lock (_processLockher)
+            {
+                if (_isStarted == false)
+                    return;
+
+                SharpProcess?.Kill();
+                ProcessExit();
+            }
         }
 
 
         public void CloseMainWindow()
         {
-            SharpProcess?.CloseMainWindow();
+            lock (_processLockher)
+            {
+                if (_isStarted == false)
+                    return;
+
+                SharpProcess?.CloseMainWindow();
+                ProcessExit();
+            }
+        }
+
+
+        private void StartWatchingForExit()
+        {
+            try
+            {
+                var handle = Handle;
+
+                if (handle == nint.Zero)
+                    return;
+
+                _waitHandle = new ProcessWaitHandle(Handle);
+                _threadPoolWaitHandle = ThreadPool.RegisterWaitForSingleObject(
+                    _waitHandle, (s, t) => ProcessExit(), _waitHandle, -1, true);
+            }
+            catch (Exception)
+            {
+
+            }
         }
 
         private string CreateCmdLine(string fileName, string cmd)
@@ -159,15 +252,18 @@ namespace StarfallAfterlife.Bridge.Native.Windows
 
         private void Dispose(bool disposing)
         {
-            if (!disposedValue)
+            if (!_disposedValue)
             {
+                if (_isStarted == true)
+                    Kill();
+
                 if (disposing)
                 {
-                    SharpProcess?.Dispose();
-                    CloseOutput();
+
                 }
 
-                disposedValue = true;
+                ReleaseUnmanagedResources();
+                _disposedValue = true;
             }
         }
 
@@ -175,5 +271,14 @@ namespace StarfallAfterlife.Bridge.Native.Windows
         {
             Dispose(disposing: true);
         }
+
+        private class ProcessWaitHandle : WaitHandle
+        {
+            public ProcessWaitHandle(IntPtr processHandle)
+            {
+                SafeWaitHandle = new SafeWaitHandle(processHandle, false);
+            }
+        }
+
     }
 }
